requirements.txt
BeautifulSoup==3.2.1
Cheetah==2.4.4
CherryPy==3.2.2
Cython==0.19.2
Flask==0.10.1
Flask-Bcrypt==0.5.2
Flask-Login==0.2.7
Flask-Mail==0.6.1
Flask-Migrate==1.2.0
Flask-OAuth==0.12
Flask-OpenID==1.1.1
## FIXME: could not find svn URL in dependency_links for this package:
Flask-SQLAlchemy==2.0-dev-20140425
Flask-Script==0.6.7
Flask-Sendmail==0.1
Flask-WTF==0.9.3
Jinja2==2.7.1
Kivy==1.7.2
Mako==0.9.1
Markdown==2.3.1
MarkupSafe==0.18
MySQL-python==1.2.4
PIL==1.1.7
Pytouch==0.1
SQLAlchemy==0.9.4
Tempita==0.5.2
Twisted==13.2.0
WTForms==1.0.5
Werkzeug==0.9.4
alembic==0.6.4
argparse==1.2.1
beautifulsoup4==4.1.3
blinker==1.3
chardet==2.1.1
decorator==3.4.0
feedparser==5.1.3
gunicorn==18.0
httplib2==0.8
itsdangerous==0.23
lamson==1.3.4
lockfile==0.9.1
matplotlib==1.2.0
mercurial==2.8
mock==1.0.1
nose==1.3.0
numpy==1.8.1
oauth==1.0.1
oauth2==1.5.211
pbr==0.8.0
procname==0.3
protobuf==2.5.0
psycopg2==2.5.2
py-bcrypt==0.4
pygame==1.9.2a0
pymongo==2.6.2
python-daemon==1.6
python-modargs==1.7
python-openid==2.2.5
pytuio==0.1
requests==0.13.2
rethinkdb==1.9.0-1
simplejson==3.3.1
six==1.6.1
sqlalchemy-migrate==0.9
stripe==1.7.9
virtualenv==1.10.1
wsgiref==0.1.2
zope.interface==4.1.1
